Project/requirements.txt
pyFirmata==1.1.0
SpeechRecognition==3.8.1
PyAudio==0.2.11
pyserial==3.5